empty-scene generator

+++ pkg/build/build.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"path/filepath"
 
+	file "github.com/bis83/basilico/pkg/file"
 	project "github.com/bis83/basilico/pkg/project"
 )
 
@@ -13,16 +14,22 @@ var fs embed.FS
 
 func Clean(baseDir string) error {
 	htmlFile := filepath.Join(baseDir, "index.html")
-	if err := os.Remove(htmlFile); err != nil {
-		return err
+	if file.Exists(htmlFile) {
+		if err := os.Remove(htmlFile); err != nil {
+			return err
+		}
 	}
 	jsFile := filepath.Join(baseDir, "app.js")
-	if err := os.Remove(jsFile); err != nil {
-		return err
+	if file.Exists(jsFile) {
+		if err := os.Remove(jsFile); err != nil {
+			return err
+		}
 	}
 	dataDir := filepath.Join(baseDir, "data")
-	if err := os.RemoveAll(dataDir); err != nil {
-		return err
+	if file.Exists(jsFile) {
+		if err := os.RemoveAll(dataDir); err != nil {
+			return err
+		}
 	}
 	return nil
 }
